Generate Identicon while parse-need happens and not on demand in the square-image itself
