+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/square-image.js
@@ -1,17 +1,11 @@
 import angular from "angular";
 
 import won from "../won-es6.js";
-import { attach, generateRgbColorArray, get, getIn } from "../utils.js";
-import { isPersona } from "../need-utils.js";
+import { attach, get, getIn } from "../utils.js";
+import * as needUtils from "../need-utils.js";
 import { connect2Redux } from "../won-utils.js";
 import { actionCreators } from "../actions/actions.js";
 import { classOnComponentRoot } from "../cstm-ng-utils.js";
-
-import Identicon from "identicon.js";
-window.Identicon4dbg = Identicon;
-
-import shajs from "sha.js";
-window.shajs4dbg = shajs;
 
 const serviceDependencies = ["$ngRedux", "$scope", "$element"];
 function genComponentConf() {
@@ -30,25 +24,16 @@ function genComponentConf() {
     constructor(/* arguments = dependency injections */) {
       attach(this, serviceDependencies, arguments);
 
-      this.personaIdenticonCacheSvg;
-      this.needIdenticonCacheSvg;
-
       const selectFromState = state => {
-        const identiconSvg = this.parseIdenticon(
-          this.uri,
-          this.needIdenticonCacheSvg
-        );
-
         const need = getIn(state, ["needs", this.uri]);
-        const personaUri = get(need, "heldBy");
+        const identiconSvg = needUtils.getIdenticonSvg(need);
 
-        const personaIdenticonSvg = this.parseIdenticon(
-          personaUri,
-          this.personaIdenticonCacheSvg
-        );
+        const personaUri = get(need, "heldBy");
+        const persona = getIn(state, ["needs", personaUri]);
+        const personaIdenticonSvg = needUtils.getIdenticonSvg(persona);
 
         return {
-          isPersona: isPersona(need),
+          isPersona: needUtils.isPersona(need),
           needInactive:
             need && get(need, "state") === won.WON.InactiveCompacted,
           needFailedToLoad:
@@ -70,29 +55,6 @@ function genComponentConf() {
       );
 
       classOnComponentRoot("won-is-persona", () => this.isPersona, this);
-    }
-
-    parseIdenticon(input, cache) {
-      if (cache) {
-        return cache;
-      }
-
-      if (!input) {
-        return;
-      }
-      // quick extra hash here as identicon.js only uses first 15
-      // chars (which aren't very unique for our uris due to the base-url):
-      const hash = new shajs.sha512().update(input).digest("hex");
-      const rgbColorArray = generateRgbColorArray(hash);
-      const idc = new Identicon(hash, {
-        size: 100,
-        foreground: [255, 255, 255, 255], // rgba white
-        background: [...rgbColorArray, 255], // rgba
-        margin: 0.2,
-        format: "svg",
-      });
-      cache = idc.toString();
-      return cache;
     }
   }
   Controller.$inject = serviceDependencies;
